Trigger topology refresh on CLUSTERDOWN
When Redis cluster topology changes quickly, we may encounter a replica
that was detached from its master and removed from the cluster. The
driver, however, may still have this replica in topology. Sending
requests to this replica trigger CLUSTERDOWN responses. To facilitate
faster recovery we may force topology request. This is exactly what this
commit does.